Create a new svelte component for the arguments
